Aula06: change SavingAccount percentage type to double

+++ src/aula06/SavingAccount.java
@@ -2,7 +2,7 @@ package aula06;
 
 public class SavingAccount  extends BankAccount{
     @Override
-    public void updateAccount(float percentage) {
+    public void updateAccount(double percentage) {
         super.updateAccount(percentage * 3);
     }
 }
